Move MessageValidator
