+++ universal-login-relayer/lib/http/relayers/Relayer.ts
@@ -25,7 +25,7 @@ import AuthorisationStore from '../../integration/sql/services/AuthorisationStor
 import WalletMasterContractService from '../../integration/ethereum/services/WalletMasterContractService';
 import {MessageStatusService} from '../../core/services/messages/MessageStatusService';
 import {SignaturesService} from '../../integration/ethereum/SignaturesService';
-import MessageValidator from '../../core/services/messages/MessageValidator';
+import MessageValidator from '../../integration/ethereum/validators/MessageValidator';
 import MessageExecutor from '../../integration/ethereum/MessageExecutor';
 import {BalanceChecker, RequiredBalanceChecker, PublicRelayerConfig} from '@universal-login/commons';
 import {DevicesStore} from '../../integration/sql/services/DevicesStore';
